requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Faker==37.1.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.1.5
MarkupSafe==3.0.2
memory-profiler==0.61.0
narwhals==1.34.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
polars==1.27.1
protobuf==5.29.4
psutil==7.0.0
py4j==0.10.9.7
pyarrow==19.0.1
pydeck==0.9.1
pymongo==4.12.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
streamlit-autorefresh==1.0.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
